refactor: enhance test infrastructure and organization
- Improve HTTPX async client mocking for better test reliability
- Refactor enhanced mocks into package structure for better organization
- Move testing documentation from backend/docs to project root docs/
- Fix test configuration to force testing environment consistently
- Enhance integration test handling with real FastAPI app fallback
- Update auth, email verification, password reset, and wallet tests
- Improve rate limiting test reliability with proper state management
